Split check out of block.
The problem here was flow control. If you follow the path of the code in
your head. You end up displaying the error message to the player before
iterating again. So unless the match is the very first card, you are
slapping their wrist. This code moves the check out of the block because
it really doesn't belong there.